+++ oap-server/server-cluster-plugin/cluster-nacos-plugin/src/main/java/org/apache/skywalking/oap/server/cluster/plugin/nacos/NacosCoordinator.java
@@ -18,28 +18,34 @@
 
 package org.apache.skywalking.oap.server.cluster.plugin.nacos;
 
+import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.api.naming.NamingService;
+import com.alibaba.nacos.api.naming.listener.Event;
+import com.alibaba.nacos.api.naming.listener.EventListener;
+import com.alibaba.nacos.api.naming.listener.NamingEvent;
 import com.alibaba.nacos.api.naming.pojo.Instance;
 import java.util.ArrayList;
 import java.util.List;
 
 import com.google.common.base.Strings;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.skywalking.oap.server.core.cluster.ClusterCoordinator;
 import org.apache.skywalking.oap.server.core.cluster.ClusterHealthStatus;
-import org.apache.skywalking.oap.server.core.cluster.ClusterNodesQuery;
-import org.apache.skywalking.oap.server.core.cluster.ClusterRegister;
 import org.apache.skywalking.oap.server.core.cluster.OAPNodeChecker;
 import org.apache.skywalking.oap.server.core.cluster.RemoteInstance;
 import org.apache.skywalking.oap.server.core.cluster.ServiceQueryException;
 import org.apache.skywalking.oap.server.core.cluster.ServiceRegisterException;
 import org.apache.skywalking.oap.server.core.remote.client.Address;
 import org.apache.skywalking.oap.server.library.module.ModuleDefineHolder;
+import org.apache.skywalking.oap.server.library.module.ModuleStartException;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
 import org.apache.skywalking.oap.server.telemetry.TelemetryModule;
 import org.apache.skywalking.oap.server.telemetry.api.HealthCheckMetrics;
 import org.apache.skywalking.oap.server.telemetry.api.MetricsCreator;
 import org.apache.skywalking.oap.server.telemetry.api.MetricsTag;
 
-public class NacosCoordinator implements ClusterRegister, ClusterNodesQuery {
+@Slf4j
+public class NacosCoordinator extends ClusterCoordinator {
 
     private final ModuleDefineHolder manager;
     private final NamingService namingService;
@@ -57,7 +63,6 @@ public class NacosCoordinator implements ClusterRegister, ClusterNodesQuery {
     public List<RemoteInstance> queryRemoteNodes() {
         List<RemoteInstance> remoteInstances = new ArrayList<>();
         try {
-            initHealthChecker();
             List<Instance> instances = namingService.selectInstances(config.getServiceName(), true);
             if (CollectionUtils.isNotEmpty(instances)) {
                 instances.forEach(instance -> {
@@ -78,6 +83,11 @@ public class NacosCoordinator implements ClusterRegister, ClusterNodesQuery {
             healthChecker.unHealth(e);
             throw new ServiceQueryException(e.getMessage());
         }
+
+        if (log.isDebugEnabled()) {
+            log.debug("Nacos cluster instances:{}", remoteInstances);
+        }
+
         return remoteInstances;
     }
 
@@ -86,17 +96,16 @@ public class NacosCoordinator implements ClusterRegister, ClusterNodesQuery {
         if (needUsingInternalAddr()) {
             remoteInstance = new RemoteInstance(new Address(config.getInternalComHost(), config.getInternalComPort(), true));
         }
+        this.selfAddress = remoteInstance.getAddress();
         String host = remoteInstance.getAddress().getHost();
         int port = remoteInstance.getAddress().getPort();
         try {
-            initHealthChecker();
             namingService.registerInstance(config.getServiceName(), host, port);
             healthChecker.health();
         } catch (Throwable e) {
             healthChecker.unHealth(e);
             throw new ServiceRegisterException(e.getMessage());
         }
-        this.selfAddress = remoteInstance.getAddress();
     }
 
     private boolean needUsingInternalAddr() {
@@ -107,6 +116,54 @@ public class NacosCoordinator implements ClusterRegister, ClusterNodesQuery {
         if (healthChecker == null) {
             MetricsCreator metricCreator = manager.find(TelemetryModule.NAME).provider().getService(MetricsCreator.class);
             healthChecker = metricCreator.createHealthCheckerGauge("cluster_nacos", MetricsTag.EMPTY_KEY, MetricsTag.EMPTY_VALUE);
+        }
+    }
+
+    private RemoteInstance buildRemoteInstance(Instance instance) {
+        Address address = new Address(instance.getIp(), instance.getPort(), false);
+        if (address.equals(selfAddress)) {
+            address.setSelf(true);
+        }
+        return new RemoteInstance(address);
+    }
+
+    private void checkHealth(List<RemoteInstance> remoteInstances) {
+        ClusterHealthStatus healthStatus = OAPNodeChecker.isHealth(remoteInstances);
+        if (healthStatus.isHealth()) {
+            this.healthChecker.health();
+        } else {
+            this.healthChecker.unHealth(healthStatus.getReason());
+        }
+    }
+
+    @Override
+    public void start() throws ModuleStartException {
+        initHealthChecker();
+        try {
+            namingService.subscribe(config.getServiceName(), new NacosEventListener());
+        } catch (NacosException e) {
+            throw new ModuleStartException("Failed to start cluster coordinator.", e);
+        }
+    }
+
+    class NacosEventListener implements EventListener {
+        @Override
+        public void onEvent(final Event event) {
+            try {
+                List<Instance> instances = ((NamingEvent) event).getInstances();
+                List<RemoteInstance> remoteInstances = new ArrayList<>(instances.size());
+                instances.forEach(instance -> {
+                    RemoteInstance remoteInstance = buildRemoteInstance(instance);
+                    if (instance.isHealthy() && instance.isEnabled()) {
+                        remoteInstances.add(remoteInstance);
+                    }
+                });
+                checkHealth(remoteInstances);
+                notifyWatchers(remoteInstances);
+            } catch (Throwable e) {
+                healthChecker.unHealth(e);
+                log.error("Failed to notify and update remote instances.", e);
+            }
         }
     }
 }